requirements.txt
backports.ssl-match-hostname==3.5.0.1
PyYAML==3.12
requests==2.12.4
six==1.10.0
websocket-client==0.40.0
